backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
aiosmtplib==2.0.0
pydantic[email]
pymysql